发票部分调整

+++ backend/src/main/java/cc/mrbird/febs/scm/dao/ScmBSupplyplanMapper.java
@@ -22,9 +22,9 @@ import java.util.List;
 public interface ScmBSupplyplanMapper extends BaseMapper<ScmBSupplyplan> {
         void updateScmBSupplyplan(ScmBSupplyplan scmBSupplyplan);
 
-        Long IsOutMenge(@Param(value = "plan") ScmBSupplyplan scmBSupplyplan);
+        Double IsOutMenge(@Param(value = "plan") ScmBSupplyplan scmBSupplyplan);
 
-        Long orderMenge(@Param(value = "plan") ScmBSupplyplan scmBSupplyplan);
+        Double orderMenge(@Param(value = "plan") ScmBSupplyplan scmBSupplyplan);
 
         int IsExistFphm(@Param(value="id") String id,@Param(value="fphm") String fphm,@Param(value="gys") String gys);
